Move event chooser logic to event
